instantiate route and service objects with additional attributes
